pyproject.toml
[project]
name = "cv-agent"
version = "0.1.0"
description = "AI-powered CV improvement system using Google ADK multi-agent workflow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.17.0",
    "google-cloud-aiplatform>=1.122.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
